requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
openai>=1.12.0
anthropic>=0.18.0
pyyaml>=6.0.1
python-multipart>=0.0.9
httpx>=0.26.0
aiofiles>=23.2.1
